murmur3.c: Fix integer conversion warnings

+++ common/murmur3.c
@@ -95,11 +95,11 @@ MurmurHash3_x86_32(const void *key, int len, uint32_t seed, void *out)
     uint32_t k1 = 0;
     switch (len & 3) {
     case 3:
-        k1 ^= tail[2] << 16;
+        k1 ^= (uint32_t)tail[2] << 16;
     case 2:
-        k1 ^= tail[1] << 8;
+        k1 ^= (uint32_t)tail[1] << 8;
     case 1:
-        k1 ^= tail[0];
+        k1 ^= (uint32_t)tail[0];
         k1 *= c1;
         k1 = ROTL32(k1, 15);
         k1 *= c2;
@@ -108,7 +108,7 @@ MurmurHash3_x86_32(const void *key, int len, uint32_t seed, void *out)
 
     //----------
     // finalization
-    h1 ^= len;
+    h1 ^= (uint32_t)len;
     h1 = fmix32(h1);
     *(uint32_t *) out = h1;
 }
@@ -164,34 +164,35 @@ FORCE_INLINE static void MurmurHash3_x86_128 (const void *key, const int len,
     uint32_t k4 = 0;
 
     switch(len & 15) {
-    case 15: k4 ^= tail[14] << 16;
-    case 14: k4 ^= tail[13] << 8;
-    case 13: k4 ^= tail[12] << 0;
+    case 15: k4 ^= (uint32_t)tail[14] << 16;
+    case 14: k4 ^= (uint32_t)tail[13] << 8;
+    case 13: k4 ^= (uint32_t)tail[12] << 0;
              k4 *= c4; k4 = ROTL32(k4,18); k4 *= c1; h4 ^= k4;
 
-    case 12: k3 ^= tail[11] << 24;
-    case 11: k3 ^= tail[10] << 16;
-    case 10: k3 ^= tail[ 9] << 8;
-    case  9: k3 ^= tail[ 8] << 0;
+    case 12: k3 ^= (uint32_t)tail[11] << 24;
+    case 11: k3 ^= (uint32_t)tail[10] << 16;
+    case 10: k3 ^= (uint32_t)tail[ 9] << 8;
+    case  9: k3 ^= (uint32_t)tail[ 8] << 0;
              k3 *= c3; k3 = ROTL32(k3,17); k3 *= c4; h3 ^= k3;
 
-    case  8: k2 ^= tail[ 7] << 24;
-    case  7: k2 ^= tail[ 6] << 16;
-    case  6: k2 ^= tail[ 5] << 8;
-    case  5: k2 ^= tail[ 4] << 0;
+    case  8: k2 ^= (uint32_t)tail[ 7] << 24;
+    case  7: k2 ^= (uint32_t)tail[ 6] << 16;
+    case  6: k2 ^= (uint32_t)tail[ 5] << 8;
+    case  5: k2 ^= (uint32_t)tail[ 4] << 0;
              k2 *= c2; k2 = ROTL32(k2,16); k2 *= c3; h2 ^= k2;
 
-    case  4: k1 ^= tail[ 3] << 24;
-    case  3: k1 ^= tail[ 2] << 16;
-    case  2: k1 ^= tail[ 1] << 8;
-    case  1: k1 ^= tail[ 0] << 0;
+    case  4: k1 ^= (uint32_t)tail[ 3] << 24;
+    case  3: k1 ^= (uint32_t)tail[ 2] << 16;
+    case  2: k1 ^= (uint32_t)tail[ 1] << 8;
+    case  1: k1 ^= (uint32_t)tail[ 0] << 0;
              k1 *= c1; k1 = ROTL32(k1,15); k1 *= c2; h1 ^= k1;
     }
 
     //----------
     // finalization
 
-    h1 ^= len; h2 ^= len; h3 ^= len; h4 ^= len;
+    h1 ^= (uint32_t)len; h2 ^= (uint32_t)len;
+    h3 ^= (uint32_t)len; h4 ^= (uint32_t)len;
 
     h1 += h2; h1 += h3; h1 += h4;
     h2 += h1; h3 += h1; h4 += h1;
@@ -212,12 +213,12 @@ FORCE_INLINE static void MurmurHash3_x86_128 (const void *key, const int len,
 
 //-----------------------------------------------------------------------------
 FORCE_INLINE static void
-MurmurHash3_x64_128(const void *key, const int len, const uint32_t seed,
+MurmurHash3_x64_128(const void *key, const size_t len, const uint32_t seed,
             void *out)
 {
     const uint8_t *data = (const uint8_t *)key;
-    const int nblocks = len / 16;
-    int i;
+    const size_t nblocks = len / 16;
+    size_t i;
     uint64_t h1 = seed;
     uint64_t h2 = seed;
     uint64_t c1 = BIG_CONSTANT(0x87c37b91114253d5);
@@ -293,8 +294,8 @@ MurmurHash3_x64_128(const void *key, const int len, const uint32_t seed,
 
     //----------
     // finalization
-    h1 ^= len;
-    h2 ^= len;
+    h1 ^= (uint64_t)len;
+    h2 ^= (uint64_t)len;
     h1 += h2;
     h2 += h1;
     h1 = fmix64(h1);
@@ -314,7 +315,7 @@ murmur3_simple(const void *keyptr)
 #ifdef __x86_64__
     uint64_t hash[2];
     MurmurHash3_x64_128(keyptr, len, 0xdeadbeef, hash);
-    return hash[1];
+    return (unsigned int)hash[1];
 #else
     if (len <= 16) {
         unsigned int hash;
